isolate meetup validation to seperate folder change test description for incorrect date add more test assertions make UI navbar fullscreen
	#
