+++ UI/assets/stylesheets/post.css
@@ -11,7 +11,7 @@ header {
   display: flex;
   position: fixed;
   top: 0;
-  width: 70%;
+  width: 100%;
   justify-content: space-evenly;
   background: #000;
   height: 10vh;
@@ -41,7 +41,7 @@ a:hover {
   display: flex;
   align-items: center;
   flex-direction: column;
-  margin-top: 70px;
+  margin-top: 15%;
 }
 
 form {
@@ -123,7 +123,7 @@ h1.head-tit {
 }
 
 .que {
-  background: #9e9e9e05;
+  background: #9e9e9e1c;
   text-align: -webkit-left;
   padding: 10px;
   border-radius: 0 10px 0 10px;
@@ -218,10 +218,6 @@ img.logout-icon:hover {
 }
 
 @media (min-width: 320px) and (max-width: 480px) {
-  header {
-    width: 100%;
-  }
-
   .app-cnt, .que-cont {
     width: 100%;
   }
