+++ routes/meetup.js
@@ -2,6 +2,8 @@ import express from 'express';
 
 import MeetupController from '../controller/MeetupController';
 
+import validateCreateMeetup from '../validations/createMeetup';
+
 const router = express.Router();
 
 router.get('/meetups', MeetupController.getAllMeetups);
@@ -10,6 +12,6 @@ router.get('/meetups/upcoming', MeetupController.upcomingMeetups);
 
 router.get('/meetups/:id', MeetupController.getSpecificMeetupRecord);
 
-router.post('/meetups', MeetupController.createMeetup);
+router.post('/meetups', validateCreateMeetup, MeetupController.createMeetup);
 
 export default router;
