+++ routes/welcome.js
@@ -6,7 +6,7 @@ const router = express.Router();
 
 // Welcome message route
 router.get('/welcome', (req, res) => {
-  const rootMessage = {
+  const data = {
     message: 'Crowd-source questions for a meetup. Questioner helps the meetup organizer prioritize questions to be answered. Other users can vote on asked questions and they bubble to the top or bottom of the log.',
     endpoints: {
       createMeetup: 'POST /api/v1/meetups',
@@ -21,7 +21,7 @@ router.get('/welcome', (req, res) => {
       createRsvp: 'POST /api/v1/meetups/:meetup_id/rsvps',
     },
   };
-  res.status(200).send({ rootMessage });
+  res.status(200).send({ data });
 });
 
 export default router;
