+++ test/question.spec.js
@@ -13,6 +13,7 @@ describe('Questions', () => {
         .end((err, res) => {
           expect(res.statusCode).to.equal(200);
           expect(res.body).to.be.an('object');
+          expect(res.body).to.have.property('data');
           done();
         });
     });
@@ -32,6 +33,7 @@ describe('Questions', () => {
         .end((err, res) => {
           expect(res.statusCode).to.equal(201);
           expect(res.body).to.be.an('object');
+          expect(res.body).to.have.property('data');
           done();
         });
     });
@@ -46,6 +48,7 @@ describe('Questions', () => {
         .end((err, res) => {
           expect(res.statusCode).to.equal(400);
           expect(res.body).to.be.an('object');
+          expect(res.body).to.have.property('message');
           expect(res.body.message).to.equal('body is required');
           done();
         });
@@ -59,6 +62,7 @@ describe('Questions', () => {
         .end((err, res) => {
           expect(res.statusCode).to.equal(400);
           expect(res.body).to.be.an('object');
+          expect(res.body).to.have.property('message');
           expect(res.body.message).to.equal('title is required');
           done();
         });
@@ -72,6 +76,7 @@ describe('Questions', () => {
         .end((err, res) => {
           expect(res.statusCode).to.equal(200);
           expect(res.body).to.be.an('object');
+          expect(res.body).to.have.property('data');
           done();
         });
     });
@@ -82,6 +87,7 @@ describe('Questions', () => {
         .end((err, res) => {
           expect(res.statusCode).to.equal(422);
           expect(res.body).to.be.an('object');
+          expect(res.body).to.have.property('message');
           expect(res.body.message).to.equal('no question with id of aaa found');
           done();
         });
@@ -95,6 +101,7 @@ describe('Questions', () => {
         .end((err, res) => {
           expect(res.statusCode).to.equal(200);
           expect(res.body).to.be.an('object');
+          expect(res.body).to.have.property('data');
           done();
         });
     });
@@ -105,6 +112,7 @@ describe('Questions', () => {
         .end((err, res) => {
           expect(res.statusCode).to.equal(422);
           expect(res.body).to.be.an('object');
+          expect(res.body).to.have.property('message');
           expect(res.body.message).to.equal('no question with id of aaa found');
           done();
         });
