+++ test/welcome.spec.js
@@ -13,6 +13,7 @@ describe('Welcome', () => {
         .end((err, res) => {
           expect(res.statusCode).to.equal(200);
           expect(res.body).to.be.an('object');
+          expect(res.body).to.have.property('data');
           done();
         });
     });
